+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes for Craft Nitro
 
+## 2.0.4 - 2021-03-04
+
+### Changed
+- Nitro will now verify and start containers if not running.
+
+### Fixed
+- Fixed an error that could occur when running `version` before `init` [#270](https://github.com/craftcms/nitro/issues/270).
+- Fixed an error that could occur when Nitro is unable to detect the database engine from a backup [#269](https://github.com/craftcms/nitro/issues/269).
+
 ## 2.0.3 - 2021-03-03
 
 ### Fixed